Delete pdacplasmaclassifier.m
Due to the use of LFS (Large File Storage) option to handle the large data (.mat) files, (to take care of the file size limits in GitHub), appropriate comments were required to be included in the code.  So that the required data files can be appropriately downloaded directly from the repository. To avoid potential confusion that may arise without those comments, the previous version of the file was deleted.